[FEAT] 닉네임수정뷰 MVVM으로 리팩터링(#65)
