+++ Plu-iOS/Plu-iOS/Scenes/NicknameEdit/Adaptor/NicknameEditNavigation.swift
@@ -1,0 +1,13 @@
+//
+//  NicknameEditNavigation.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/21.
+//
+
+import Foundation
+
+protocol NicknameEditNavigation {
+    func backButtonTapped()
+    func nicknameChangeCompleteButtonTapped()
+}
