+++ Plu-iOS/Plu-iOS/Scenes/NicknameEdit/Model/NicknameEditInput.swift
@@ -1,0 +1,15 @@
+//
+//  NicknameEditInput.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/21.
+//
+
+import Foundation
+import Combine
+
+struct NicknameEditInput {
+    let textFieldSubject: AnyPublisher<String, Never>
+    let naviagtionLeftButtonTapped: PassthroughSubject<Void, Never>
+    let naviagtionRightButtonTapped: PassthroughSubject<String?, Never>
+}
