+++ Plu-iOS/Plu-iOS/Scenes/NicknameEdit/ViewModel/NicknameEditViewModel.swift
@@ -1,0 +1,10 @@
+//
+//  NicknameViewModel.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/21.
+//
+
+import Foundation
+
+protocol NicknameEditViewModel: ViewModel where Input == NicknameEditInput, Output == NicknameEditOutput {}
