+++ Plu-iOS/Plu-iOS/Scenes/Onboarding/Model/OnboardingInput.swift
@@ -1,0 +1,15 @@
+//
+//  OnboardingInput.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/21.
+//
+
+import Foundation
+import Combine
+
+struct OnboardingInput {
+    let textFieldSubject: AnyPublisher<String, Never>
+    let backButtonTapped: PassthroughSubject<Void, Never>
+    let singInButtonTapped: PassthroughSubject<String, Never>
+}
