+++ Plu-iOS/Plu-iOS/Scenes/Onboarding/Model/OnboardingOutput.swift
@@ -1,0 +1,14 @@
+//
+//  OnboardingOutput.swift
+//  Plu-iOS
+//
+//  Created by uiskim on 2023/12/21.
+//
+
+import Foundation
+import Combine
+
+struct OnboardingOutput {
+    let nickNameResultPublisher: AnyPublisher<NicknameState, Never>
+    let signInStatePublisher: AnyPublisher<LoadingState, Never>
+}
