+++ Plu-iOS/Plu-iOS/Scenes/Onboarding/ViewModel/OnboardingViewModel.swift
@@ -6,19 +6,6 @@
 //
 
 import Foundation
-import Combine
 
-struct OnboardingInput {
-    let textFieldSubject: AnyPublisher<String, Never>
-    let backButtonTapped: PassthroughSubject<Void, Never>
-    let singInButtonTapped: PassthroughSubject<String, Never>
-}
-
-struct OnboardingOutput {
-    let nickNameResultPublisher: AnyPublisher<NicknameState, Never>
-    let signInStatePublisher: AnyPublisher<LoadingState, Never>
-}
-
-protocol OnboardingViewModel: ViewModel where Input == OnboardingInput, Output == OnboardingOutput {
-}
+protocol OnboardingViewModel: ViewModel where Input == OnboardingInput, Output == OnboardingOutput {}
